removing reach and merging select and multiple
