+++ libs/pages/application/src/lib/ui/page-settings-general/page-settings-general.spec.tsx
@@ -2,7 +2,6 @@ import { getByTestId, screen, waitFor } from '@testing-library/react'
 import { render } from '__tests__/utils/setup-jest'
 import { wrapWithReactHookForm } from '__tests__/utils/wrap-with-react-hook-form'
 import { BuildModeEnum, BuildPackLanguageEnum, GitProviderEnum } from 'qovery-typescript-axios'
-import { upperCaseFirstLetter } from '@console/shared/utils'
 import PageSettingsGeneral, { PageSettingsGeneralProps } from './page-settings-general'
 
 describe('PageSettingsGeneral', () => {
@@ -34,12 +33,9 @@ describe('PageSettingsGeneral', () => {
       })
     )
 
-    const buildMode = screen.getByTestId('input-select-mode')
-    const dockerfile = screen.getByTestId('input-text-dockerfile')
-
     screen.getByDisplayValue('hello-world')
-    expect(buildMode.querySelector('.input__value')?.textContent).toContain(upperCaseFirstLetter(BuildModeEnum.DOCKER))
-    expect(dockerfile.getAttribute('value')).toBe('Dockerfile')
+    screen.getByText('Docker')
+    screen.getByDisplayValue('Dockerfile')
   })
 
   it('should render the form with buildpack section', () => {
@@ -51,16 +47,9 @@ describe('PageSettingsGeneral', () => {
       })
     )
 
-    const buildMode = screen.getByTestId('input-select-mode')
-    const language = screen.getByTestId('input-select-language')
-
     screen.getByDisplayValue('hello-world')
-    expect(buildMode.querySelector('.input__value')?.textContent).toContain(
-      upperCaseFirstLetter(BuildModeEnum.BUILDPACKS)
-    )
-    expect(language.querySelector('.input__value')?.textContent).toContain(
-      upperCaseFirstLetter(BuildPackLanguageEnum.CLOJURE)
-    )
+    screen.getByText('Buildpacks')
+    screen.getByText('Clojure')
   })
 
   it('should submit the form', async () => {
