+++ libs/shared/ui/src/index.ts
@@ -13,7 +13,6 @@ export * from './lib/components/warning-screen-mobile/warning-screen-mobile'
 export * from './lib/components/icon/icon'
 export * from './lib/components/icon-fa/icon-fa'
 export * from './lib/components/icon/icon-awesome.enum'
-export * from './lib/components/inputs/input-select/input-select'
 export * from './lib/components/inputs/input-text/input-text'
 export * from './lib/components/inputs/input-text-area/input-text-area'
 export * from './lib/components/inputs/input-toggle/input-toggle'
@@ -69,5 +68,5 @@ export * from './lib/components/avatar/avatar'
 export * from './lib/components/link/link'
 export * from './lib/components/help-section/help-section'
 export * from './lib/components/truncate/truncate'
-export * from './lib/components/inputs/input-select-multiple/input-select-multiple'
+export * from './lib/components/inputs/input-select/input-select'
 export * from './lib/components/table/table-row-deployment/table-row-deployment'
