+++ libs/shared/ui/src/lib/components/inputs/input-select/input-select.stories.tsx
@@ -1,4 +1,5 @@
 import { Meta, Story } from '@storybook/react'
+import { Value } from '@console/shared/interfaces'
 import { InputSelect, InputSelectProps } from './input-select'
 
 export default {
@@ -6,23 +7,34 @@ export default {
   title: 'Inputs/InputSelect',
 } as Meta
 
+const options: Value[] = [
+  {
+    label: 'Test 1',
+    value: 'test1',
+  },
+  {
+    label: 'Test 2',
+    value: 'test2',
+  },
+  {
+    label: 'Test 3',
+    value: 'test3',
+  },
+]
+
 const Template: Story<InputSelectProps> = (args) => <InputSelect {...args} />
 
-export const Primary = Template.bind({})
-Primary.args = {
-  label: 'Type of use',
-  items: [
-    {
-      label: 'Personal',
-      value: 'personal',
-    },
-    {
-      label: 'Work',
-      value: 'work',
-    },
-    {
-      label: 'School',
-      value: 'school',
-    },
-  ],
+export const Single = Template.bind({})
+Single.args = {
+  label: 'Select',
+  options: options,
+  value: 'test1',
+}
+
+export const Multi = Template.bind({})
+Multi.args = {
+  label: 'Select Multiple',
+  options: options,
+  isMulti: true,
+  value: ['test1', 'test2'],
 }
